release version 1.0.4

+++ src/config/index.ts
@@ -4,14 +4,14 @@ export const DOMAIN_BASE_CONFIG = {
 }
 
 export const EIP712_NONCE_TYPES = {
-  Nonce: [{ name: 'nonce', type: 'uint' }],
+  Nonce: [{ name: 'nonce', type: 'uint256' }],
 }
 
 export const EIP712_OFFCHAIN_TRANSACTION_TYPES = {
   OffChainTransaction: [
     { name: 'to', type: 'address' },
-    { name: 'value', type: 'uint' },
+    { name: 'value', type: 'uint256' },
     { name: 'data', type: 'bytes' },
-    { name: 'nonce', type: 'uint' },
+    { name: 'nonce', type: 'uint256' },
   ],
 }
